SwitchOpt determined by AggressiveIntTypeSpec breaks repeated String cases
This change fixes an issue where the function GlobOpt::IsSwitchOptEnabled was updated to add conditions
for type specialization. This resulted in bad IR being produced, which impacted how type-specialized BailOuts
were constructed and which registers were restored (or, not restored). The fix is to apply align optimization
checks in SwitchIRBuilder based on the types of the cases.

CR Feedback fixes