Disable a number of OACR warnings whose signal-to-noise ratio is far too low
These warnings typically involve conditional expressions and Boolean
operators -- things like  <expr> && 1, <expr> || 1, 'if' condition known to
be true, etc.  These can indicate problems, but more frequently for us, we
do these intentionally.  (For example, we might have <expr1> || MACRO(...),
which can lead to this warning if MACRO is defined to be a compile-time
constant in release builds.)