Module : memory leak fix when the syntax error happens.
We were no removing modulerecord from MessageQueue when the exception happens - fixed that.
We were printing the exception in the notifymoduleready callback, which is not needed.
Populating the description field in the case of error, so that correct error will be propagated in the event of exception (sometime we use message field, other times description).