Add timeout for stackoverflow test
When a stackoverflow happens within a try finally enclosed inside try catch, we have to do a stackwalk till we find the try catch frame, this means walking all the frames that caused the stackoverflow.
This test takes longer than the default timeout value on xplat. Moving it to slow test.