Some small improvements on IO error message formatting
1. Adds FORMAT_MESSAGE_IGNORE_INSERTS flag when we call FormatMessage(),
so that if some IO error messages contain insert sequences, it won't crash
the program due to null dereference.

2. If FormatMessage() is called with FORMAT_MESSAGE_ALLOCATE_BUFFER, we need
to free the buffer. It doesn't really matter because the process exits immediately
after printing the error, but it adds noise to analysis tools.

3. To check if TTDHostMKDir() succeeds, I think it's better to check if errno == EEXIST,
because errno can have values other than EEXIST and ENOENT, such as EACCES (access denied).