OS#9030969 fix issues with stack traces for inlined functions
It is possible for inlinee functions to have statements at the address as the call site. In such cases the stack trace would pick up the statement number of the inlinee's statement instead of the statement associated with the inliner's call to the inlinee. The statement associated with the inliner's call always occurs before the address of the call, so this PR makes the stack walker look for the first statement before the inlinee start, and ensures that that statement gets recorded.