[CVE-2018-0874] - Chakra Array Includes Uninitialized Memory RCE - Individual
Callbacks while sorting causes the head of an array to not have any missing value, although later we cause an exception in the sorting. Which left the array
in the inconsistent state. Later in the HeadSegmentIndexOfHelper we exploit that situation.
Fixed that by reseting the no-missing value state in the no-exception case. And also put fail-fast where we don't expect it happen.