Fix #2949: Class const name binding incorrectly marked as module export storage
We were mistakenly marking both the class let and const decls as module export storage. This resulted in empty instantiated scopes and incorrect class behavior in some cases.

The fix is not to mark the const binding as module export storage, as it is visible to the inside of the class only.

Added some tests to verify the bindings of exported classes.