[CVE-2018-0933] Chakra: JIT - Incomplete Fix for MSRC-41913 - Google, Inc.
This change addresses a scenario where a deepCopy of a native array is needed when its head segment is already on the heap. In this case, it bypasses the previous fix because the head is on the stack and thus fails to do a deepCopy.
The fix is to unconditionally reallocate both the array object and its segments when deepCopy is true.