reinstate LdLen in bytecode with additional profile info
Array optimizations in the backend require precise information about the type of the array being used. The FldInfo profile data associated with LdFld opcodes contains only the type of the value being loaded, so bailouts occuring on array type checks for LdLen didn't always result in updated profiled data, which could lead to infinite bailouts.

This commit reintroduces LdLen to the bytecode generator, but changes the way profile data is handled for it. LdLen now has profile data split into 2 places. It has the   FldInfo associated with a LdFld instruction to allow object.length to be optimized, but also has a separate LdLenInfo entry containing the type of the object being loaded from. This choice of implementation was driven by the fact that FldInfo in the profile data uses the same index as the InlineCache.

LdLen_A is given a OpLayoutDynamicProfile<ElementCP> layout to include both inline cache index and the ProfileId for the LdLenInfo. In the interpreter LdLen_A is handled the same as LdFld except for the addition of recording the object type being loaded from.