Hoist speculation masking when possible
This change attempts to move speculation-related masking from the
inside of loops to the out-edges.

The bulk of the work happens in the dead store pass; I took advantage
of the multiple passes on loops to gather data about speculated uses,
and then the final dead store pass handles marking instructions which
we think are safe and storing the data for the Lowerer.

There's a new set of data structures introduced in ClusterList.h; the
purpose of these is to try to get relatively efficient set joins when
dealing with the sets of symbols here. There's a couple of parts that
may be a little tailored for this use case (segmentclusterlist is set
up for symbol clustering patterns and for two-pass runs), but I think
we might find another use for them sometime.

The Lowering code is pretty limited to handling the new opcode, which
has a list of symbols to mask, and expands to instructions that block
speculation on those symbols.