Silence a warning raised due to ETW generated header
A pre-defined function is added to the chakra etw header that does
not initialize a buffer before passing it around. It looks ok, but
the warning is in a file that we have limited control over. Here I
disabled the warning for the whole included file, which seems like
the closest we can scope it.