Speed up `in` operations with PropertyString or Symbol keys, part 1
This change allows the runtime call for `str in obj` to use the same property record cache that we use for loading `obj[str]`. This makes the `in` operator faster in some common cases, and also populates the cache for an upcoming change which will lower fast-path code for `in` on PropertyStrings and Symbols.