Allow naming labels and highlighting instructions
This change is designed to make following some code paths easier when
doing JIT work. The first change is that it adds debug-only code that
puts names on labels, which should help users looking at instructions
that expand to a great number of blocks. The second change is to have
developers be able to set a color for a particular Instr such that in
dumps it stands out better, which may be useful in some situations.