Path dependent branch folding
This optimization redirects branches to their eventual destination, when we can prove the result of all indermediate branches, and is legal to skip all the intermediate instructions.
Example :

```
function foo(x) {
  var testValue = 100;
  if (x > 100) {
    testValue = 0;  // At this block we can directly redirect the branch to return block, instead of doing following test again
  }
  if (testValue != 0) {
    print(testValue);
  }

```
This helps in efficient merge of valueInfos, so that out global optimizer can aggressively work on blocks, for which it was previously limited due to conservative merge of valueInfos.
This is the first part of the optimization, a second implementation supporting additional instructions will be next.

fix prefast warning

CR fixes

Fix reorg missing case

More CR fixes

Add correct bytecodeoffset