Fix potential buffer overflow in FNonZeroFiniteDblToStr
For the function Js::NumberUtilities::FNonZeroFiniteDblToStr (in NumberUtilities_strtod.cpp),
if the input double value dbl is negative and nDstBufSize is zero, there's a buffer overflow.
The buffer length is not checked before the buffer ppsz is written.

This is not currently triggerable, since all callers pass a 256 byte buffer.