Initialize ICU when initializing Intl in order to catch data file issues early
If ICU can't load the data file, it will report back U_FILE_ACCESS_ERROR or U_MISSING_RESOURCE_ERROR on calls to ICU APIs that require the data file. When using Windows ICU, the only reason we wouldn't be able to load the data file is if we are out of memory (with bundled ICU, the data should be statically or dynamically linked, and this shouldn't be an issue). As such, to reduce noise from ICU_ASSERT, we can load the data file as soon as we initialize Intl using u_init, and if that fails with a known error code, we can treat it as OOM.