[CVE-2018-8128] Edge - UAF of borrowed inline caches after redeferral
Do not allow ScriptFunctionWithInlineCache to borrow its InlineCache* from the FunctionBody, because in certain cases of redeferral this will cause jitted code to access stale pointers. Instead of borrowing the caches from the FunctionBody, create a base ScriptFunction and let the runtime access the current FunctionBody inline caches when the function is executed.