Use consistent error messages in invalid `extends` clause
A class that tries to extend from a number might get two separate error messages depending on whether the number happens to be represented internally as a JavascriptNumber heap object or a tagged int. Either behavior is fine by [the spec](https://tc39.github.io/ecma262/#sec-runtime-semantics-classdefinitionevaluation), but we should be consistent to avoid failures in tools that use string comparison.

Fixes OS:17285098