Moving HomeObj field out of ScriptFunction
Only class member functions and object literal methods have home object set. So moving this field out of ScriptFunction should help us reduce the size of normal ScriptFunctions.

As part of this the SetHomeObj opcode is also removed. Instead we set the home object along with the creation of the function.