OS#17436985: Fix stack overflow in SimpleDataCacheWrapper::SeekReadStreamToBlockHelper
It's possible for IStream::Read to return S_FALSE without actually reading any bytes from the stream. In that case we will infinitely recurse looking for a valid block because S_FALSE is not a failed HRESULT. To make sure we don't hit this, check to see if we failed to read the block type and bail early. Also convert S_FALSE into E_FAIL in the read and write functions of SimpleDataCacheWrapper to make it fail earlier.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/17436985