pack dictionary entries better
We frequently have recycler entries that are one pointer and two ints, but they take up 24 bytes instead of 16 due to inefficient packing. With this change, we can ask the compiler to put the `next` integer wherever leads to a smaller representation: either after the value or after the key.

This change also simplifies the `Clear` method on dictionary entries to use `TValue()` or `TKey()` to reset integral values to zero. A few more complicated types still require specializations.