MSFT#17472203 - Data stream underlying SimpleDataCacheWrapper may return S_OK when bytesRead < bytesExpected
The stream we read data from in the SimpleDataCacheWrapper can read fewer bytes than we requested and still return S_OK. We expect to get back S_FALSE in this case but that doesn't always happen. Fix the wrapper to assert we didn't overflow and convert S_FALSE into a failing hr.

Also added debug tracing to SimpleDataCacheWrapper operations and disable generating and serializing the parser state cache for non-cacheable resources.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/17472203