Store ByteCodeCache in the deferred stubs during bytecode serialization
The deferred stubs store a cache pointer which is used to find the serialized string table and lookup serialized names for the set of captured names. We allocate this cache pointer (which is a wrapper over the raw serialized bytecode buffer) during bytecode deserialization. That means for the parser state cache, we have to serialize bytecode into the cache and wait for the next load when we can recover the buffer bytes from the cache or else we would need to serialize bytecode and immediately deserialize it again. Both solutions are not ideal so this change causes the bytecode serializer to create this cache pointer and store it in deferred stubs during serialization in parser state generation mode. The deferred stubs may now be used on the first load.