Add ScopeInfo to serialized bytecode for deferred functions
When we undefer a defer-parsed function, we need the ScopeInfo in order to rehydrate the scope chain above that function. We never previously serialized that data but now that we can serialize and deserialize deferred functions, we need to also save the ScopeInfo.

This is implemented by keeping a map of LocalFunctionId to FunctionInfo* in the ByteCodeCache used by the bytecode reader. We need to keep this in the ByteCodeCache (arena allocated) because defer-deserialized functions may need this map when they are deserialized.