Enable Deferred Parsing for Background Parse
This change turns on Deferred Parsing in BGParseManager, which will reduce the amount of time needed to parse in the background. Appropriate flags are added in calls inside BGParseManager and BGParseWorkItem.
Also, in order to serialize defer parsed bytecode, ScriptContext::CompileUTF8Core refactors the Parser object outside of the call because the serializer needs state from the parser to complete.