OS#17542375 Correct possible overflow of deferred stubs array
When undeferring a function with deferred stubs, we try and use those stubs to determine if expressions beginning with a left paren are nested lambda functions. We do this when we see a left paren and the next deferred stub is a lambda function starting at the same character. Unfortunately we don't keep track of the count of deferred stubs in the current deferred stubs array. If all of the nested functions in the function we're undefering are located in source before the left paren character, we should not check for the next deferred stub as this would overflow the array and cause a possible AV.

Fix this by tracking the count of stubs in the current deferred stubs array.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/17542375