SubString crash
I recently updated LiteralStringWithPropertyStringPtr to refer to the PropertyRecord's copy of the string buffer (just like PropertyString does), which can save some memory. However, if the script then takes a substring of that string, it would keep an invalid pointer as `originalFullStringReference`, which could allow the buffer to be freed.

The fix contains three parts:

1. Return the correct owning pointer from `LiteralStringWithPropertyStringPtr::GetOriginalStringReference`.
2. Add an assertion that would catch this bug eagerly, rather than only after a subsequent GC.
3. Update `SingleCharString::GetOriginalStringReference` to pass the new assertion. It was not a bug because single-char strings are pinned to a script context and copied upon marshaling, but this change makes its behavior more consistent with the other string types.

Fixes OS:17362430