OS#16855035 - Properly support redeferral for serialized defer-parsed functions
When we redeferred and subsequently reparsed a function which was defer-parsed and deserialized from bytecode, we took the null captured names set off of the new ParseNode and used it to generate a set of empty deferred stubs for the function. When we skipped scanning nested functions we didn<C2><92>t push references to their captured names. This caused us to try and load from invalid slots at runtime. We also lost the set of deferred stubs stored in ParseableFunctionInfo AuxPtrs during one of the duplication hops when we duplicate a FunctionProxy.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/16855035