Fix assertion when dumping encoder with foreground JIT
While outputting offsets in the EncoderPhase, several asserts can be hit because the state of the EntryPointInfo is Queued, where it is expected to be Recorded or Done. The effect is that there is no base address available for the right output.
The fix is to move the calls with that dependency to after the state transitions to Recorded.