OS#17529874 - Skipping a nested deferred function containing a with statement can trigger load from invalid scope slot
When we use deferred stubs in the parser to skip past nested deferred functions, we need to be careful about nested functions containing with statements. Names referenced inside the with can't be pushed as ordinary references - we need to know that the symbols are referenced inside the with statement or not. This isn't possible with stubs currently so, instead, avoid skipping such nested functions.

Fixes:
https://microsoft.visualstudio.com/_workitems/edit/17529874