OS#17530048 - AssertMsg(i < this->Length(), "index out of bound") -- Chakra!BVFixed::AssertRange
Function count saved in the serialized bytecode is computed based on the total number of functions we visit and add to the bytecode. It is possible, though, for us to have a higher function id after defer parsing due to the reparse scenario we have for lambda argument lists. We avoid lambda parameter reparse when using the parser state cache except in the presence of default arguments. So serializing and running bytecode containing a lambda function with a default argument containing another function will increment the next function id to be greater than the size of the startup functions array in SourceDynamicProfileManager which will hit this assert.

Example:
```js
(a = () => 1) => 2;
(function(){})(); // <-- function id of this function is equal to the size of startupFunctions array
```

Fix by serializing the function count from the source info instead of counting the functions we add to the bytecode one-by-one.

Fixes:
https://microsoft.visualstudio.com/_workitems/edit/17530048
https://microsoft.visualstudio.com/_workitems/edit/17529917