Use more efficient call to load module
The ThreadProc for BackgroundJobProcessor calls GetModuleHandleEx to ensure that Chakra.dll isn't unloaded while code is still executing on this thread. However, using the full path to the DLL is an expensive kernal call.
Instead, follow the convention of other threads in Chakra and use GET_MODULE_HANDLE_EX_FLAG_FROM_ADDRESS.