OS#18260560 - ASSERTION : scope at GetEnclosingFunc
Function expression with nested-function declared in default arguments containing a with statement which maybe escapes a nested function causes assert.

```javascript
(function foo(a = function bar() {
  with ({}) {
      foo;
  }
}()) {})();
```

We try and look at the param scope and body scope but we don't check the function expression scope in `ByteCodeGenerator::CheckDeferParseHasMaybeEscapedNestedFunc`. Simple fix is to check function expression scope if param and body scope are nullptr.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/18260560

Found vis oss-fuzz