Support deferred stubs for functions declared in parameter scope
We previously skipped creating deferred function stubs for functions declared in the parameter scope. This throws-off the order of nested functions under functions with children defined in the parameter scope. We worked-around this by not using deferred stubs in parents with default arguments.

Add support for creating deferred stubs and using them when parsing function parameter lists.