[CVE-2018-8367] Edge - UAF for Edge on WIP - Qihoo 360
This change fixes an issue where deepCopied arrays have a different inlined head segment behavior from its original instance. This is because deepCopy'ing an array would unconditionally allocate the head segment as inline, regardless of its size. This conflicts with JavascriptArray::HasInlineHeadSegment, which is based upon size rather than layout. In the PoC, this causes problems when removing elements via Slice.
The fix is to clone the inlined behavior of the original instance in addition to the instance's data. This allows related invariants to be maintained.

This change also ensures that the aligned, allocated size of the inline head segment on the stack is never larger than INLINE_CHUNK_SIZE, similarly to NewLiteral.