oopjit branch fold BrTrue and BrFalse for true, false, and numbers.
allows the branch from ```if (a)``` to be folded when ```a``` is known to be true, false, or a number. This will also improve the IsIn optimization which currently replaces the instr with LdTrue when it is can be proven, and will improve any potential wins of allowing bools to be fixed fields.